_XEditResGet32: Fix casts to avoid unexpected sign extension in 64-bit
Like _XEditResGet16 we need to cast to our destination type, not
the source type, before combining the upper & lower halves.

+++ src/EditresCom.c
@@ -1884,7 +1884,7 @@ _XEditResGet32(ProtocolStream *stream, unsigned long *value)
     if (!(_XEditResGet16(stream, &temp1) && _XEditResGet16(stream, &temp2)))
 	return (False);
 
-    *value = ((unsigned short)temp1 << (XER_NBBY * 2)) + (unsigned short)temp2;
+    *value = ((unsigned long)temp1 << (XER_NBBY * 2)) + (unsigned long)temp2;
     return (True);
 }
